Add support for cloning mame to all monitors
